Fix and refactor styles
