--- src/client/src/containers/PageAddServices/ServiceGroup/messages.ts
@@ -1,13 +1,0 @@
-import { defineMessages } from "react-intl";
-
-const messages = defineMessages({
-  hostingTitle: {
-    id: "PageAddServices.serviceGroup.title",
-    defaultMessage: "Publish your project to the web",
-  },
-  storageTitle: {
-    id: "PageAddServices.serviceGroup.title",
-    defaultMessage: "Store your data in the cloud",
-  },
-});
-export default messages;
