+++ src/client/src/containers/PageAddServices/ServiceGroup/styles.module.css
@@ -5,12 +5,12 @@
   padding: 0em 0em;
 }
 
-.serviceType {
+.title {
   font-size: 1.2em;
   font-weight: 600;
 }
 
-.descriptor {
+.description {
   font-size: 2em;
 }
